requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-telegram-bot==21.0
python-dotenv==1.0.0
telethon==1.36.0
gunicorn==21.2.0
